requirements.txt
Babel==2.15.0
certifi==2024.6.2
charset-normalizer==3.3.2
glfw==2.7.0
icalendar==5.0.13
idna==3.7
numpy==2.0.0
pybind11==2.12.0
PyOpenGL==3.1.7
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
skia-python==87.6
urllib3==2.2.2
